+++ test/server/transpiler.test.ts
@@ -3,14 +3,12 @@ import { once } from "node:events";
 import fs, { constants as fsConstants } from "node:fs";
 import { EOL } from "node:os";
 
-import { usingTemporaryFiles } from "using-temporary-files";
-
 import { Transpiler } from "../../src/server/transpiler.js";
+import { withTemporaryFiles } from "../lib/with-temporary-files.js";
 
 const TYPESCRIPT_SOURCE = `export const x:number = 1;${EOL}`;
 const JAVASCRIPT_SOURCE = `export const x = 1;${EOL}`;
 const JAVASCRIPT_SOURCE_COMMONJS = `"use strict";${EOL}Object.defineProperty(exports, "__esModule", { value: true });${EOL}exports.x = void 0;${EOL}exports.x = 1;${EOL}`;
-
 describe("a Transpiler", () => {
   let transpiler: Transpiler = new Transpiler("src", "dist", "");
 
@@ -19,20 +17,21 @@ describe("a Transpiler", () => {
   });
 
   it("finds a file and transpiles it", async () => {
-    await usingTemporaryFiles(async ($) => {
-      await $.add("src/found.ts", TYPESCRIPT_SOURCE);
-      transpiler = new Transpiler($.path("src"), $.path("dist"), "module");
+    const files = {
+      "src/found.ts": TYPESCRIPT_SOURCE,
+    };
+
+    await withTemporaryFiles(files, async (basePath, { path }) => {
+      transpiler = new Transpiler(path("src"), path("dist"), "module");
 
       await transpiler.watch();
 
       // eslint-disable-next-line promise/avoid-new, no-promise-executor-return
       await new Promise((resolve) => setTimeout(resolve, 1000));
 
-      expect(fs.readdirSync($.path("./dist"))).toEqual(["found.js"]);
+      expect(fs.existsSync(path("dist/found.js"))).toBe(true);
 
-      expect(fs.existsSync($.path("dist/found.js"))).toBe(true);
-
-      expect(fs.readFileSync($.path("dist/found.js"), "utf8")).toBe(
+      expect(fs.readFileSync(path("dist/found.js"), "utf8")).toBe(
         JAVASCRIPT_SOURCE,
       );
 
@@ -43,46 +42,43 @@ describe("a Transpiler", () => {
   it("discovers a new file and transpiles it", async () => {
     // on Linux the watcher doesn't seem to work consistently if there's not a file in the directory to begin with
 
-    await usingTemporaryFiles(async ($) => {
-      await $.add("src/starter.ts", TYPESCRIPT_SOURCE);
-      transpiler = new Transpiler($.path("src"), $.path("dist"), "module");
+    await withTemporaryFiles(
+      { "src/starter.ts": TYPESCRIPT_SOURCE },
+      async (basePath, { add, path }) => {
+        transpiler = new Transpiler(path("src"), path("dist"), "module");
 
-      await transpiler.watch();
+        await transpiler.watch();
 
-      const write = once(transpiler, "write");
-      const error = once(transpiler, "error");
+        const write = once(transpiler, "write");
+        const error = once(transpiler, "error");
 
-      await $.add("src/added.ts", TYPESCRIPT_SOURCE);
+        await add("src/added.ts", TYPESCRIPT_SOURCE);
 
-      if (process.platform === "win32") {
-        // Chokidar's add event seems to be unreliable on Windows
-        // Not sure what to do about it, so just skip this test
-        return;
-      }
+        if (process.platform === "win32") {
+          // Chokidar's add event seems to be unreliable on Windows
+          // Not sure what to do about it, so just skip this test
+          return;
+        }
 
-      await Promise.race([write, error]);
+        await Promise.race([write, error]);
 
-      expect(fs.readdirSync($.path("./dist"))).toEqual([
-        "added.js",
-        "starter.js",
-      ]);
+        expect(fs.readFileSync(path("dist/added.js"), "utf8")).toBe(
+          JAVASCRIPT_SOURCE,
+        );
 
-      expect(fs.readFileSync($.path("dist/added.js"), "utf8")).toBe(
-        JAVASCRIPT_SOURCE,
-      );
-
-      await transpiler.stopWatching();
-    });
+        await transpiler.stopWatching();
+      },
+    );
   });
 
   it("sees an updated file and transpiles it", async () => {
-    await usingTemporaryFiles(async ($) => {
-      transpiler = new Transpiler($.path("src"), $.path("dist"), "module");
+    const files = {
+      "src/update-me.ts": "const x = 'code to be overwritten';\n",
+    };
 
-      await $.add(
-        "src/update-me.ts",
-        `const x = 'code to be overwritten';${EOL}`,
-      );
+    await withTemporaryFiles(files, async (basePath, { add, path }) => {
+      transpiler = new Transpiler(path("src"), path("dist"), "module");
+
       const initialWrite = once(transpiler, "write");
 
       await transpiler.watch();
@@ -90,12 +86,10 @@ describe("a Transpiler", () => {
 
       const overwrite = once(transpiler, "write");
 
-      await $.add("src/update-me.ts", TYPESCRIPT_SOURCE);
+      await add("src/update-me.ts", TYPESCRIPT_SOURCE);
       await overwrite;
 
-      expect(fs.readdirSync($.path("./dist"))).toEqual(["update-me.js"]);
-
-      expect(fs.readFileSync($.path("dist/update-me.js"), "utf8")).toBe(
+      expect(fs.readFileSync(path("dist/update-me.js"), "utf8")).toBe(
         JAVASCRIPT_SOURCE,
       );
 
@@ -104,18 +98,19 @@ describe("a Transpiler", () => {
   }, 10_000);
 
   it("sees a removed TypeScript file and deletes the JavaScript file", async () => {
-    await usingTemporaryFiles(async ($) => {
-      await $.add("src/delete-me.ts", TYPESCRIPT_SOURCE);
-      transpiler = new Transpiler($.path("src"), $.path("dist"), "module");
+    const files = {
+      "src/delete-me.ts": TYPESCRIPT_SOURCE,
+    };
+
+    await withTemporaryFiles(files, async (basePath, { path, remove }) => {
+      transpiler = new Transpiler(path("src"), path("dist"), "module");
 
       await transpiler.watch();
-      await $.remove("src/delete-me.ts");
+      await remove("src/delete-me.ts");
       await once(transpiler, "delete");
 
-      expect(fs.readdirSync($.path("./dist"))).toEqual([]);
-
       expect(() => {
-        fs.accessSync($.path("dist/delete-me.js"), fsConstants.F_OK);
+        fs.accessSync(path("dist/delete-me.js"), fsConstants.F_OK);
       }).toThrow(/ENOENT/u);
 
       await transpiler.stopWatching();
@@ -123,43 +118,23 @@ describe("a Transpiler", () => {
   });
 
   it("transpiles as CommonJS when specified", async () => {
-    await usingTemporaryFiles(async ($) => {
-      await $.add("src/found.ts", TYPESCRIPT_SOURCE);
+    const files = {
+      "src/found.ts": TYPESCRIPT_SOURCE,
+    };
 
-      transpiler = new Transpiler($.path("src"), $.path("dist"), "commonjs");
-
-      await transpiler.watch();
-
-      expect(fs.readdirSync($.path("./dist"))).toEqual(["found.cjs"]);
-
-      expect(fs.existsSync($.path("dist/found.cjs"))).toBe(true);
-
-      expect(fs.readFileSync($.path("dist/found.cjs"), "utf8")).toBe(
-        JAVASCRIPT_SOURCE_COMMONJS,
-      );
-
-      await transpiler.stopWatching();
-    });
-  });
-
-  it("converts requires of .js files to .cjs", async () => {
-    await usingTemporaryFiles(async ($) => {
-      await $.add(
-        "src/importer.ts",
-        'import local from "./local.js"; local();',
-      );
-
-      transpiler = new Transpiler($.path("src"), $.path("dist"), "commonjs");
+    await withTemporaryFiles(files, async (basePath, { path }) => {
+      transpiler = new Transpiler(path("src"), path("dist"), "commonjs");
 
       await transpiler.watch();
-      expect(fs.readdirSync($.path("./dist"))).toEqual(["importer.cjs"]);
 
       // eslint-disable-next-line promise/avoid-new, no-promise-executor-return
       await new Promise((resolve) => setTimeout(resolve, 1000));
 
-      const contents = fs.readFileSync($.path("dist/importer.cjs"));
+      expect(fs.existsSync(path("dist/found.cjs"))).toBe(true);
 
-      expect(contents.includes('require("./local.cjs")')).toBe(true);
+      expect(fs.readFileSync(path("dist/found.cjs"), "utf8")).toBe(
+        JAVASCRIPT_SOURCE_COMMONJS,
+      );
 
       await transpiler.stopWatching();
     });
